基础配置

+++ src/main/java/com/hrbc/business/service/impl/CandidatesServiceImpl.java
@@ -39,7 +39,7 @@ public class CandidatesServiceImpl implements CandidatesService {
     @Override
     public PageResultDTO loadPage(PageQueryParamDTO params) {
 
-        int count = 0;
+        long count = 0;
         List<Candidates> list = null;
         int page = 1;
         int size = 10;
